Added bloom, thanks to DrTwox for the bloom example code.

+++ Gravity/source/FPS/fps.h
@@ -12,4 +12,4 @@
 #endif /* FPS_H_ */
 
 u8 calculateFPS();
-int profiler(int start);
+u64 profiler(int start);
